+++ src/store/modules/tagsViewRoutes.ts
@@ -1,7 +1,4 @@
 import { Module } from 'vuex';
-import { ElMessage } from 'element-plus';
-import screenfull from 'screenfull';
-import { Local } from '/@/utils/storage';
 // 此处加上 `.ts` 后缀报错，具体原因不详
 import { TagsViewRoutesState, RootStateTypes } from '/@/store/interface/index';
 
@@ -9,7 +6,7 @@ const tagsViewRoutesModule: Module<TagsViewRoutesState, RootStateTypes> = {
 	namespaced: true,
 	state: {
 		tagsViewRoutes: [],
-		isCurrenFullscreen: false,
+		isTagsViewCurrenFull: false,
 	},
 	mutations: {
 		// 设置 TagsView 路由
@@ -17,8 +14,8 @@ const tagsViewRoutesModule: Module<TagsViewRoutesState, RootStateTypes> = {
 			state.tagsViewRoutes = data;
 		},
 		// 设置卡片全屏
-		getCurrenFullscreen(state: any, data: boolean) {
-			state.isCurrenFullscreen = data;
+		getCurrenFullscreen(state: any, bool: boolean) {
+			state.isTagsViewCurrenFull = bool;
 		},
 	},
 	actions: {
@@ -27,34 +24,8 @@ const tagsViewRoutesModule: Module<TagsViewRoutesState, RootStateTypes> = {
 			commit('getTagsViewRoutes', data);
 		},
 		// 设置卡片全屏
-		setCurrenFullscreen({ commit }, timeout: number = 0) {
-			const screenfulls: any = screenfull;
-			if (!screenfulls.isEnabled) {
-				ElMessage.warning('暂不不支持全屏');
-				return false;
-			}
-			setTimeout(() => {
-				const currenFullscreenChange = () => {
-					const layoutViewBgWhite = document.querySelector('.layout-view-bg-white') as HTMLElement;
-					if (screenfulls.isFullscreen) {
-						commit('getCurrenFullscreen', true);
-						// 设置全屏时，设置有 `layout-view-bg-white` 类的高度
-						if (layoutViewBgWhite) layoutViewBgWhite.style.height = `calc(100vh - 30px)`;
-					} else {
-						screenfulls.off('change', currenFullscreenChange);
-						commit('getCurrenFullscreen', false);
-						if (!layoutViewBgWhite) return false;
-						const getThemeConfig = Local.get('themeConfig');
-						if (getThemeConfig) {
-							let { isTagsview } = getThemeConfig;
-							if (isTagsview) layoutViewBgWhite.style.height = `calc(100vh - 114px)`;
-							else layoutViewBgWhite.style.height = `calc(100vh - 80px)`;
-						}
-					}
-				};
-				screenfulls.on('change', currenFullscreenChange);
-				screenfulls.toggle();
-			}, timeout);
+		setCurrenFullscreen({ commit }, bool: Boolean) {
+			commit('getCurrenFullscreen', bool);
 		},
 	},
 };