+++ src/views/login/index.vue
@@ -120,7 +120,7 @@ export default {
 			position: absolute;
 			top: 0;
 			right: 0;
-			width: 51px;
+			width: 50px;
 			height: 50px;
 			overflow: hidden;
 			cursor: pointer;
@@ -134,19 +134,22 @@ export default {
 				border-bottom: 50px solid #ffffff;
 				border-right: 50px solid transparent;
 				z-index: 2;
+				top: 0;
+				right: 0;
 			}
 			&:hover {
 				opacity: 1;
 				transition: all ease 0.3s;
+				color: var(--color-primary);
 			}
 			i {
 				content: '';
-				width: 50px;
+				width: 48px;
 				height: 50px;
 				position: absolute;
-				top: -2px;
+				top: 0px;
 				right: 0px;
-				font-size: 50px;
+				font-size: 47px;
 				z-index: 1;
 			}
 		}