+++ src/views/params/dynamic/index.vue
@@ -1,9 +1,10 @@
 <template>
 	<div class="layout-view-bg-white flex" :style="{ height: `calc(100vh - ${setViewHeight}` }">
 		<div class="flex-margin">
-			<el-result icon="warning" title="动态路由" subTitle="请根据提示进行操作">
+			<el-result icon="warning" title="动态路由" subTitle="可 `开启 TagsView 共用` 进行单标签测试">
 				<template #extra>
-					<el-button type="primary" size="small" icon="iconfont icon-dongtai" @click="onGoDetailsClick">点击测试动态路由传参</el-button>
+					<el-input v-model="value" placeholder="请输入路由参数id值" clearable></el-input>
+					<el-button type="primary" size="small" icon="iconfont icon-dongtai" class="mt15" @click="onGoDetailsClick">动态路由传参</el-button>
 				</template>
 			</el-result>
 		</div>
@@ -18,7 +19,9 @@ export default defineComponent({
 	name: 'paramsDynamic',
 	setup() {
 		const store = useStore();
-		const state = reactive({});
+		const state = reactive({
+			value: '',
+		});
 		const router = useRouter();
 		// 设置 view 的高度
 		const setViewHeight = computed(() => {
@@ -29,15 +32,14 @@ export default defineComponent({
 		// 跳转到详情
 		const onGoDetailsClick = () => {
 			// name 值为路由中的 name
-			const t = new Date().getTime();
-			const id = Math.random();
 			router.push({
 				name: 'paramsDynamicDetails',
 				params: {
-					t,
-					id,
+					t: 'vue-next-admin',
+					id: state.value,
 				},
 			});
+			state.value = '';
 		};
 		return {
 			setViewHeight,